Started on Issue 17
Created a vocab table and updated activities accordingly, but there is
an error when the function selectFromRange is called, having to do with
the cursor. Once that is resolved and the activity works as expected,
the lag still needs to be fixed, and the new activity needs to be
created.