Implemented searchAll() in SearchActivity
Added a Result class to track matches for a general search. It is now
possible to retrieve kanji and vocab words through a search term without
specifying what is being searched. An improvement that should be made in
the future is to handle similar spellings of words- to retrieve items
without an exact match.